requirements.txt
gefpy==0.5.4.4
geojson==2.5.0
glog==0.3.1
h5py==3.3.0
matplotlib==3.4.2
numexpr==2.7.3
numpy==1.21.0
pandas==1.2.5
Pillow==9.1.0
scipy==1.7.0
seaborn==0.11.1
skimage==0.0
tifffile==2022.4.26
torch==1.9.0
